+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Enforce JWT on API side
 - Only return results by user ID
+- Replace file input with selects in sample creation
 
 ### Removed
 - Remove author_email field as it is redundant and slows down the user